+++ packages/template-retail-react-app/CHANGELOG.md
@@ -9,7 +9,7 @@
 - Product scroller: don't skip tiles if window is too large [#2003](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2003)
 - PDP / PLP: Render non HTTP 404 erros [#2003](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2003)
 - Error page: Render home page when clicking nav icon [#2003](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2003)
-- Encode non ASCII HTTP headers when `ENCODE_NON_ASCII_HTTP_HEADERS` flag is set to true in `constants.js` [#2009](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2009)
+- Encode non ASCII HTTP headers when `encodeNonAsciiHttpHeaders` flag is set to true in `ssr.js` [#2009](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2009)
 
 ### Accessibility Improvements
 - [a11y] Hide svg from screenreader as they are decorative on homepage [#1980](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1980)